requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
branca==0.8.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
folium==0.20.0
fonttools==4.61.0
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.53.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jieba==0.42.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.12.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
RapidFuzz==3.14.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
soupsieve==2.8
SQLAlchemy==2.0.44
streamlit==1.51.0
streamlit-folium==0.25.3
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
wordcloud==1.9.4
xyzservices==2025.11.0
